gem: add json gem to gemfile
The provided ruby package in the Arch Linux Extra repository (currently 3.0.6-1) is missing the json gem for some reason. This may not be an issue in the package found within the Extra staging repository (currently 3.2.4-1), however, I have not tested it and most users should NOT be using packages from the staging repositories.