gemfile: fix site building on ruby 3.4
csv and base64 gems were removed from the ruby 3.4 standard library, this prevents the site from building. Similarly, the logger gem is set to be removed from the ruby 3.5.0 standard library. This commit adds these gems to the gemfile to prevent warnings and allows the site to continue to build on newer ruby versions