css: explicitly set light.css to load when perfers-color-scheme is light
Light CSS has an override where some icons are explicitly overriden,
however the dark CSS does not have this override.

Without explicitly specifying that light CSS should only be loaded in
light mode, this causes this little overridden CSS to fallthrough to
dark CSS when the theme is set to default, causing carousel arrows to
be black in dark mode.

This doesn't happen when a theme is explicitly set because on those
cases the CSS is loaded directly from the respective CSS file rather
than the default style.css selector.